refactor: extract route handler
