+++ internal/sbi/routes.go
@@ -1,0 +1,26 @@
+package sbi
+
+import "github.com/gin-gonic/gin"
+
+type Route struct {
+	Method  string
+	Pattern string
+	APIFunc gin.HandlerFunc
+}
+
+func applyRoutes(group *gin.RouterGroup, routes []Route) {
+	for _, route := range routes {
+		switch route.Method {
+		case "GET":
+			group.GET(route.Pattern, route.APIFunc)
+		case "POST":
+			group.POST(route.Pattern, route.APIFunc)
+		case "PUT":
+			group.PUT(route.Pattern, route.APIFunc)
+		case "PATCH":
+			group.PATCH(route.Pattern, route.APIFunc)
+		case "DELETE":
+			group.DELETE(route.Pattern, route.APIFunc)
+		}
+	}
+}
